requirements.txt
jupyterlab
pyprojroot
tqdm
ipywidgets
haystack-ai
fastembed
lancedb
tantivy
fastembed-haystack
duckdb